caldera_orchestrator_project/requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
rich>=13.7.0
fpdf2>=2.7.9
